refactor: status 에 관한 훅을 추출한다
