+++ src/pages/Lines/hooks.js
@@ -1,0 +1,9 @@
+import { useSelector } from "react-redux";
+import useStatus from "../../hooks/useStatus";
+import { fetchLines, reset, selectLinesStatus } from "./slice";
+
+export const useLinesStatus = () => {
+  const status = useSelector(selectLinesStatus);
+
+  return useStatus({ status, fetch: fetchLines, reset });
+};
